Refactor bot structure

+++ app/database/models/users.py
@@ -4,6 +4,6 @@ from app.database import db
 class User(db.Model):
     __tablename__ = "users"
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BIGINT, primary_key=True)
     name = db.Column(db.Text)
     email = db.Column(db.Text)
